Add rule: users can't delete other users' events

+++ src/com/timeline/project/server/DeleteEventServlet.java
@@ -9,6 +9,7 @@ import javax.servlet.http.HttpServletResponse;
 import com.google.appengine.api.datastore.DatastoreService;
 import com.google.appengine.api.datastore.DatastoreServiceFactory;
 import com.google.appengine.api.datastore.Entity;
+import com.google.appengine.api.datastore.EntityNotFoundException;
 import com.google.appengine.api.datastore.Key;
 import com.google.appengine.api.datastore.KeyFactory;
 import com.google.appengine.api.users.User;
@@ -28,13 +29,25 @@ public class DeleteEventServlet extends HttpServlet {
         
         String eventId = req.getParameter("id");
         
-        Entity entity = new Entity("Event");
         if (eventId != null) {
         	System.out.println("Deleting id " + eventId);
         	Key eventKey = KeyFactory.createKey("Event", eventId);
         	DatastoreService datastore = DatastoreServiceFactory.getDatastoreService();
         	
-        	datastore.delete(eventKey);
+        	Entity event = null;
+        	try {
+				event = datastore.get(eventKey);
+			} catch (EntityNotFoundException e) {
+				System.out.println("Event does not exist. Not deleted.");
+				return;
+			}
+        	
+        	String uId = (String) event.getProperty("userId");
+        	if (uId == userId) {
+        		datastore.delete(eventKey);        		
+        	} else {
+        		System.out.println("Current user does not match requested user event. Not deleted.");
+        	}
         }
         else
         {
